+++ src/Devel/Config.hs
@@ -16,7 +16,7 @@ Closest thing is in the stack Main module.
         
 Will be rewritten to depend on the stack library.
 -}
-
+{-# LANGUAGE ScopedTypeVariables #-}
 
 module Devel.Config 
 ( setConfig
@@ -25,8 +25,12 @@ module Devel.Config
 ) where
 
 import System.Process (readProcessWithExitCode)
-import System.Environment (unsetEnv, setEnv)
-import System.Exit (ExitCode(..))
+import System.Environment (unsetEnv, setEnv, lookupEnv)
+import Control.Exception (catch)
+
+import System.FilePath.Posix (pathSeparator)
+import System.Directory (getCurrentDirectory, getHomeDirectory, doesDirectoryExist)
+import System.Info (arch, compilerName, os)
 
 import Devel.Types
 
@@ -37,22 +41,75 @@ setConfig = do
   _ <- unsetEnv "GHC_PACKAGE_PATH"
   _ <- setEnv "PATH" path
   setEnv "GHC_PACKAGE_PATH" pkgDb
+  
 
 getConfig :: IO Config
 getConfig = do
-  (exitCode, stdout, stderr) <- readProcessWithExitCode "stack" ["path"] ""
-  case exitCode of
-    ExitSuccess   -> parseConfig stdout
-    ExitFailure _ -> fail stderr
+  -- If stack isn't installed use cabal-install.
+  catch getStackConfig getCabalConfig
   where 
+    getStackConfig :: IO Config
+    getStackConfig = do 
+      (_, stdout, _) <- readProcessWithExitCode "stack" ["path"] ""
+      parseConfig stdout
+
+
     parseConfig :: String -> IO Config
     parseConfig stdout = do
       let outputList = lines stdout
           tupleList = map (span (/=':') ) outputList
           path = concatMap getPath tupleList
-          pkgDb = concatMap getPkgDb tupleList ++ ":"
+          pkgDb = concatMap getPkgDb tupleList
       return (path, pkgDb)
 
+getCabalConfig :: IOError -> IO Config
+getCabalConfig _ = do
+  mPath <- lookupEnv "PATH"
+  mPkgDb <- lookupEnv "GHC_PACKAGE_PATH"
+  
+  ghcVersion <- catch getGhcVersion getGhcVersionError
+  
+  cwd <- getCurrentDirectory
+  homeDir <- getHomeDirectory
+  isSandBoxed <- doesDirectoryExist ".cabal-sandbox"
+  
+  let compilerVersion' = last $ words ghcVersion
+      forSandboxPath = arch ++ "-" ++ os ++ "-" ++ compilerName ++ "-" ++ compilerVersion'
+      forUserPath    = arch ++ "-" ++ os ++ "-" ++ compilerVersion'
+      forGlobalPath  = compilerName ++ "-" ++ compilerVersion'
+
+
+      sandboxPkgDb = if isSandBoxed
+                        then cwd ++ (pathSeparator : ".cabal-sandbox") 
+                             ++ (pathSeparator : forSandboxPath) ++ "-packages.conf.d:"
+                        else "" -- We're not using a sandbox.
+
+      userPkgDb = homeDir ++ (pathSeparator: ".ghc") ++ (pathSeparator:forUserPath) 
+                  ++ (pathSeparator:"package.conf.d:")
+      globalPkgDb = "/usr/lib/" ++ forGlobalPath ++"/package.conf.d:"
+
+      pkgDb' = sandboxPkgDb ++ userPkgDb ++ globalPkgDb
+
+      path = case mPath of
+                   Just p -> p
+                   Nothing -> fail "The environment variable PATH isn't set."
+      pkgDb = case mPkgDb of
+                   Just db -> db
+                   Nothing -> pkgDb'
+
+  return (path, pkgDb)
+
+getGhcVersion :: IO String
+getGhcVersion = do
+  (_, stdout, _) <- readProcessWithExitCode "ghc" ["--version"] ""
+  return stdout
+
+-- Give proper fail information.
+getGhcVersionError :: IOError -> IO String
+getGhcVersionError _ = do
+  fail $ "Is GHC not in your PATH? "
+         ++ "Because wai-devel can't get the version number from: "
+         ++ "`ghc --version`"
 
 getPath :: (String, String) -> String
 getPath (key,value)
@@ -61,5 +118,5 @@ getPath (key,value)
 
 getPkgDb :: (String, String) -> String
 getPkgDb (key,value)
-          | key == "snapshot-pkg-db" = dropWhile (==':') $ filter (/=' ') value
+          | key == "ghc-package-path" = dropWhile (==':') $ filter (/=' ') value
           | otherwise = ""